python_backend/requirements.txt
Flask>=2.0
SQLAlchemy>=1.4
requests>=2.25
Flask-HTTPAuth>=4.4
python-dotenv>=0.19
Flask-SQLAlchemy>=2.5
Flask-CORS>=3.0 # Added for Cross-Origin Resource Sharing
gunicorn>=20.1 # Added WSGI server for better performance